ENH Transform into a Jugfile
Merge steps 1 and 2 into a single Jugfile
